Прога 5 за 3 дня (успех?), день 2/3 Что я пишу и я ли вообще?

+++ Прога/Лабораторная 5/lab5/Lab5/app/src/main/java/ru.itmo.prog.lab5/utility/Element.java
@@ -1,5 +1,13 @@
 package ru.itmo.prog.lab5.utility;
 
+/**
+ * Абстрактный класс для элементов коллекции, поддерживающих уникальный идентификатор и сравнение.
+ * @author pmih
+ */
 public abstract class Element implements Comparable<Element>, Validatable {
-    abstract public int getId();
+    /**
+     * Возвращает уникальный идентификатор элемента.
+     * @return Идентификатор в формате long.
+     */
+    abstract public long getId();
 }
